requirements.txt
langgraph>=0.2.26
langchain>=0.2.12
langchain-core>=0.2.12
langchain-google-genai>=1.0.6
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
jinja2>=3.1.4

